Revert "Solve not being able to retrieve the AbstractProver in ProverWithAssumptionsWrapper by moving it (and its subclasses) up 1 package into the same package as AbstractProver + get the nested AbstractProver from other delegates if needed" Due tthe solution that was worked towards was adding a lot of complexity that we most likely don't need. We will first try to work around this, but maybe come back to this later. This reverts commit bff027debd568b7774d6d70c4df58485a2f4f207.
