+++ src/org/sosy_lab/java_smt/basicimpl/withAssumptionsWrapper/BasicProverWithAssumptionsWrapper.java
@@ -1,17 +1,12 @@
-/*
- * This file is part of JavaSMT,
- * an API wrapper for a collection of SMT solvers:
- * https://github.com/sosy-lab/java-smt
- *
- * SPDX-FileCopyrightText: 2026 Dirk Beyer <https://www.sosy-lab.org>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
+// This file is part of JavaSMT,
+// an API wrapper for a collection of SMT solvers:
+// https://github.com/sosy-lab/java-smt
+//
+// SPDX-FileCopyrightText: 2020 Dirk Beyer <https://www.sosy-lab.org>
+//
+// SPDX-License-Identifier: Apache-2.0
 
-package org.sosy_lab.java_smt.basicimpl;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkNotNull;
+package org.sosy_lab.java_smt.basicimpl.withAssumptionsWrapper;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -31,12 +26,7 @@ class BasicProverWithAssumptionsWrapper<T, P extends BasicProverEnvironment<T>>
   protected final List<BooleanFormula> solverAssumptionsAsFormula = new ArrayList<>();
 
   BasicProverWithAssumptionsWrapper(P pDelegate) {
-    // We may need to access the AbstractProver in a nested fashion if P is also a delegate
-    checkArgument(
-        pDelegate instanceof AbstractProver<?>
-            || pDelegate instanceof InterpolatingProverDelegate<?>
-            || pDelegate instanceof OptimizationProverDelegate);
-    delegate = checkNotNull(pDelegate);
+    delegate = pDelegate;
   }
 
   protected void clearAssumptions() {
@@ -75,12 +65,10 @@ class BasicProverWithAssumptionsWrapper<T, P extends BasicProverEnvironment<T>>
     return delegate.isUnsat();
   }
 
-  @SuppressWarnings("resource")
   @Override
   public boolean isUnsatWithAssumptions(Collection<BooleanFormula> assumptions)
       throws SolverException, InterruptedException {
     clearAssumptions();
-    getDelegateAsAbstractProver().checkIsUnsatOverAssumptions(assumptions);
     solverAssumptionsAsFormula.addAll(assumptions);
     // Since we are using the delegates isUnsat() impl, we don't need to update
     // wasLastSatCheckSatisfiable etc.
@@ -141,18 +129,5 @@ class BasicProverWithAssumptionsWrapper<T, P extends BasicProverEnvironment<T>>
       throws InterruptedException, SolverException {
     clearAssumptions();
     return delegate.allSat(pCallback, pImportant);
-  }
-
-  @SuppressWarnings("unchecked")
-  private AbstractProver<?> getDelegateAsAbstractProver() {
-    if (delegate instanceof AbstractProver<?> pAbstractProverDelegate) {
-      return pAbstractProverDelegate;
-    } else if (delegate instanceof InterpolatingProverDelegate<?> pInterpolatingProverDelegate) {
-      return pInterpolatingProverDelegate.getDelegateAsAbstractProver();
-    } else if (delegate instanceof OptimizationProverDelegate pOptimizationProverDelegate) {
-      return pOptimizationProverDelegate.getDelegateAsAbstractProver();
-    }
-    // Should never happen due to the check in the constructor
-    throw new IllegalStateException();
   }
 }
