+++ src/org/sosy_lab/java_smt/basicimpl/withAssumptionsWrapper/InterpolatingProverWithAssumptionsWrapper.java
@@ -1,14 +1,12 @@
-/*
- * This file is part of JavaSMT,
- * an API wrapper for a collection of SMT solvers:
- * https://github.com/sosy-lab/java-smt
- *
- * SPDX-FileCopyrightText: 2026 Dirk Beyer <https://www.sosy-lab.org>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
+// This file is part of JavaSMT,
+// an API wrapper for a collection of SMT solvers:
+// https://github.com/sosy-lab/java-smt
+//
+// SPDX-FileCopyrightText: 2020 Dirk Beyer <https://www.sosy-lab.org>
+//
+// SPDX-License-Identifier: Apache-2.0
 
-package org.sosy_lab.java_smt.basicimpl;
+package org.sosy_lab.java_smt.basicimpl.withAssumptionsWrapper;
 
 import static com.google.common.base.Preconditions.checkNotNull;
