+++ src/org/sosy_lab/java_smt/test/InterpolatingProverWithAssumptionsWrapperTest.java
@@ -9,7 +9,7 @@
 package org.sosy_lab.java_smt.test;
 
 import org.sosy_lab.java_smt.api.InterpolatingProverEnvironment;
-import org.sosy_lab.java_smt.basicimpl.InterpolatingProverWithAssumptionsWrapper;
+import org.sosy_lab.java_smt.basicimpl.withAssumptionsWrapper.InterpolatingProverWithAssumptionsWrapper;
 import org.sosy_lab.java_smt.solvers.opensmt.Logics;
 
 public class InterpolatingProverWithAssumptionsWrapperTest
